requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-telegram-bot
requests
